+++ test/security_token_registry.js
@@ -667,7 +667,6 @@ contract('SecurityTokenRegistry', accounts => {
                     bytesSTO,
                     (1000 * Math.pow(10, 18)),
                     (1000 * Math.pow(10, 18)),
-                    true,
                     {
                         from: token_owner,
                         gas: 26000000
@@ -697,7 +696,6 @@ contract('SecurityTokenRegistry', accounts => {
                 bytesSTO,
                 (1000 * Math.pow(10, 18)),
                 (1000 * Math.pow(10, 18)),
-                true,
                 {
                     from: token_owner,
                     gas: 2500000